merging the Call code with freedcamp issues

+++ Chatalyze/helper/extension/DateStamp.swift
@@ -10,11 +10,13 @@ import Foundation
 import UIKit
 
 public extension Date {
+    
+    
 
-    var millisecondsSince1970:Int {
+    var millisecondsSince1970:Int64 {
 
         let interval = Int64((self.timeIntervalSince1970 * 1000.0).rounded())
-        return Int(interval)
+        return interval
         //return Int(self.timeIntervalSince1970)
     }
